fw/uno: split deterministic ECDSA-P384 into crypto/ecc_det.rs
Move the RFC 6979 deterministic-sign path out of the ~1080-line
crypto/ecc.rs into a dedicated crypto/ecc_det.rs, so the base HsmEcc
trait adapter stays focused on the standard curve operations. The new
module will also host deterministic ECC key generation later.

Moved: the P-384 order/base-point constants, montgomery_size,
be_sub_assign, secure_zero, the RFC 6979 HMAC-SHA384 DRBG, and the
ecc_sign_with_k / ecc_generate_k_rfc6979 / ecc_sign_deterministic
methods (plus the throwaway KAT self-tests). PRIME384_LE is shared with
the base HsmEcc impl, so it stays in ecc.rs and is now pub(super).

No functional change.

Copilot-Session: 2444329e-909e-4ced-b5df-5a40fb65bb6c
